+++ Qwerty.ECS.Runtime/Chunks/EcsChunk.cs
@@ -32,12 +32,12 @@ namespace Qwerty.ECS.Runtime.Chunks
         
         internal unsafe T ReadComponent<T>(int index, int typeIndex) where T : struct, IEcsComponent
         {
-            return MemoryUtil.Read<T>(body, header->rowSizeInBytes * index + header->ReadOffsetByType((short)typeIndex));
+            return MemoryUtil.Read<T>(body, header->rowSizeInBytes * index + header->ReadOffsetByType(typeIndex));
         }
 
         internal unsafe void WriteComponent<T>(int index, int typeIndex, T component) where T : struct, IEcsComponent
         {
-            MemoryUtil.Write(body, header->rowSizeInBytes * index + header->ReadOffsetByType((short)typeIndex), component);
+            MemoryUtil.Write(body, header->rowSizeInBytes * index + header->ReadOffsetByType(typeIndex), component);
         }
 
         internal unsafe void Dispose()
