+++ Qwerty.ECS.Runtime/Components/EcsComponentTypeHandle.cs
@@ -3,8 +3,8 @@ namespace Qwerty.ECS.Runtime.Components
 {
     public readonly struct EcsComponentTypeHandle<T> where T : struct, IEcsComponent
     {
-        public readonly short typeIndex;
-        internal EcsComponentTypeHandle(short typeIndex)
+        public readonly int typeIndex;
+        internal EcsComponentTypeHandle(int typeIndex)
         {
             this.typeIndex = typeIndex;
         }
